% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play the prediction game.}
\usage{
play(n, guess_1, guess_2, FUN)
}
\arguments{
\item{n}{The number of times to play the game.}

\item{guess_1}{A number.}

\item{guess_2}{A number.}

\item{FUN}{A formula to apply to data. Must begin with ~ and may include
references to data that are in the global workspace.}
}
\value{
A tibble with the results of the game.
}
\description{
Play the prediction game.
}
\examples{
play(n = 5, guess_1 = 3, guess_2 = 7, ~ sample(x = 1:10, size = 1))
}
