% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play the prediction game.}
\usage{
play(n, guess_1, guess_2, formula)
}
\arguments{
\item{n}{The number of times to play the game.}

\item{guess_1}{A number.}

\item{guess_2}{A number.}

\item{formula}{A formula to apply to data. Must begin with ~ and may include
references to data that are in the global workspace.}
}
\value{
A tibble with the results of the game.
}
\description{
Play the prediction game.
}
\examples{
z <- play(n = 5, guess_1 = 3, guess_2 = 7, ~ sample(x = 1:10, size = 1))
z
z <- play(n = 500, guess_1 = 20.09, guess_2 = 21, 
          ~ mean(sample(mtcars$mpg, size = 10, replace = FALSE)))
table(z$winner)
}
