% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play the prediction game.}
\usage{
play(data, n, guess_1, guess_2)
}
\arguments{
\item{data}{A numeric vector.}

\item{n}{The number of times to play the game.}

\item{guess_1}{A number.}

\item{guess_2}{A number.}
}
\value{
A tibble with the results of the game.
}
\description{
Play the prediction game.
}
\examples{
play(1:10, 5, 3, 7)
}
