% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{Play the prediction game.}
\usage{
play(data, n, guess_1, guess_2, FUN, ...)
}
\arguments{
\item{data}{A numeric vector.}

\item{n}{The number of times to play the game.}

\item{guess_1}{A number.}

\item{guess_2}{A number.}

\item{FUN}{A function to apply to data.}

\item{...}{Optional arguments to FUN.}
}
\value{
A tibble with the results of the game.
}
\description{
Play the prediction game.
}
\examples{
play(data = 1:10, n = 5, 3, 7, sample, size = 1)
}
